# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/app/admin/system/seed-demo:
    post:
      operationId: AdminController_seedDemo
      parameters: []
      responses:
        '200':
          description: Summary record creati per tabella
      security:
        - access-token: []
      summary: Carica i dati demo (SuperAdmin only)
      tags:
        - Admin
  /api/v1/app/admin/system/reset-demo:
    post:
      operationId: AdminController_resetDemo
      parameters: []
      responses:
        '200':
          description: Summary record cancellati per tabella
      security:
        - access-token: []
      summary: >-
        Cancella tutti i dati tranne il SuperAdmin (SuperAdmin only,
        irreversibile)
      tags:
        - Admin
  /api/v1/app/admin/users:
    get:
      operationId: AdminController_findAllUsers
      parameters:
        - name: p
          required: false
          in: query
          description: Pagina (default 1)
          schema:
            type: string
        - name: i
          required: false
          in: query
          description: Elementi per pagina (default 20)
          schema:
            type: string
        - name: q
          required: false
          in: query
          description: Ricerca testuale su nome, cognome, email
          schema:
            type: string
        - name: o
          required: false
          in: query
          description: Campo di ordinamento (name, surname, email, role, createdAt)
          schema:
            type: string
        - name: d
          required: false
          in: query
          description: Direzione ordinamento (asc|desc, default asc)
          schema:
            type: string
      responses:
        '200':
          description: Lista utenti paginata
      security:
        - access-token: []
      summary: Lista utenti paginata (un Admin non vede gli utenti SuperAdmin)
      tags:
        - Admin
    post:
      operationId: AdminController_createUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: Utente creato
        '403':
          description: Tentativo di creare un utente SuperAdmin da parte dell'Admin
      security:
        - access-token: []
      summary: Crea un nuovo utente (un Admin non può creare utenti SuperAdmin)
      tags:
        - Admin
  /api/v1/app/admin/users/{guid}:
    get:
      operationId: AdminController_findOneUser
      parameters:
        - name: guid
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Utente trovato
        '403':
          description: Target SuperAdmin non gestibile dall'Admin
        '404':
          description: Utente non trovato
      security:
        - access-token: []
      summary: Dettaglio utente (un Admin non vede gli utenti SuperAdmin)
      tags:
        - Admin
    patch:
      operationId: AdminController_updateUser
      parameters:
        - name: guid
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: Utente aggiornato
        '403':
          description: Target SuperAdmin non gestibile dall'Admin
        '404':
          description: Utente non trovato
      security:
        - access-token: []
      summary: Aggiorna i dati di un utente
      tags:
        - Admin
  /api/v1/app/admin/users/{guid}/toggle-active:
    patch:
      operationId: AdminController_toggleActiveUser
      parameters:
        - name: guid
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Stato isActive invertito
        '403':
          description: Target SuperAdmin non gestibile dall'Admin
        '404':
          description: Utente non trovato
      security:
        - access-token: []
      summary: Abilita/disabilita un utente senza eliminarlo
      tags:
        - Admin
  /api/v1/app/admin/users/{guid}/reset-mfa:
    post:
      operationId: AdminController_resetMfaUser
      parameters:
        - name: guid
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: MFA resettata con successo
        '403':
          description: Target SuperAdmin non gestibile dall'Admin
        '404':
          description: Utente non trovato
      security:
        - access-token: []
      summary: Resetta l'MFA di un utente (potrà ri-configurarla al prossimo login)
      tags:
        - Admin
  /api/v1/app/admin/audit-log:
    get:
      operationId: AdminController_findAuditLog
      parameters:
        - name: p
          required: false
          in: query
          description: Pagina (default 1)
          schema:
            type: string
        - name: i
          required: false
          in: query
          description: Elementi per pagina (default 20)
          schema:
            type: string
        - name: userId
          required: false
          in: query
          description: Filtra per id dell'utente autore
          schema:
            type: string
        - name: action
          required: false
          in: query
          description: Filtra per azione (ricerca parziale, es. "login")
          schema:
            type: string
        - name: from
          required: false
          in: query
          description: Data/ora inizio (ISO 8601)
          schema:
            type: string
        - name: to
          required: false
          in: query
          description: Data/ora fine (ISO 8601)
          schema:
            type: string
      responses:
        '200':
          description: Lista audit log paginata
      security:
        - access-token: []
      summary: Lista paginata degli eventi di audit log (Admin+)
      tags:
        - Admin
...
info:
  title: Starter Kit API
  description: >-
    API REST del boilerplate aziendale (auth, RBAC, MFA, audit log, gestione
    utenti).
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateUserDto:
      type: object
      properties:
        name:
          type: string
          description: Nome dell'utente
          example: Mario
        surname:
          type: string
          description: Cognome dell'utente
          example: Rossi
        email:
          type: string
          description: Email dell'utente
          example: mario.rossi@example.com
        role:
          type: number
          description: Ruolo assegnato
          enum:
            - 5
            - 10
            - 20
            - 30
          example: 30
        scopeId:
          type: string
          description: >-
            Identificatore di scope multi-tenant/multi-sede (libero per il
            progetto che eredita lo starter-kit)
      required:
        - name
        - email
        - role
    UpdateUserDto:
      type: object
      properties:
        name:
          type: string
          description: Nome dell'utente
        surname:
          type: string
          description: Cognome dell'utente
        email:
          type: string
          description: Email dell'utente
        role:
          type: number
          description: Ruolo assegnato
          enum:
            - 5
            - 10
            - 20
            - 30
        scopeId:
          type: string
          description: Identificatore di scope multi-tenant/multi-sede
...